+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [22.6.0] - Aspose Words Cloud for Dart 22.6 Release Notes
+
+- Added 'DeleteBookmark' and 'DeleteBookmarkOnline' API methods for delete bookmarks by name from the document.
+- Added 'DeleteBookmarks' and 'DeleteBookmarksOnline' API methods for delete all bookmarks from the document.
+- Added 'InsertBookmark' and 'InsertBookmarkOnline' API methods for create new bookmarks in the document.
+- Support all save formats for 'CreateDocument' operation.
+
+
+## [22.5.0] - Aspose Words Cloud for Dart 22.5 Release Notes
+
+- Internal API fixes and improvments.
+
+
 ## [22.4.0] - Aspose Words Cloud for Dart 22.4 Release Notes
 
 - Added ExportShapesAsSvg to HtmlSaveOption.